refactor: use userGroupRepository to get resource permission userGroups
