+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2025-01-15T17:33:26.984Z\n"
-"PO-Revision-Date: 2025-01-15T17:33:26.984Z\n"
+"POT-Creation-Date: 2025-01-20T08:19:05.880Z\n"
+"PO-Revision-Date: 2025-01-20T08:19:05.880Z\n"
 
 msgid "Low"
 msgstr ""
