+++ src/CompositionRoot.ts
@@ -83,9 +83,6 @@ import { DeleteResourceFileUseCase } from "./domain/usecases/DeleteResourceFileU
 import { ResourceFileTestRepository } from "./data/repositories/test/ResourceFileTestRepository";
 import { ResourceFileRepository } from "./domain/repositories/ResourceFileRepository";
 import { ResourceFileD2Repository } from "./data/repositories/ResourceFileD2Repository";
-import { ResourcePermissionsD2Repository } from "./data/repositories/ResourcePermissionsD2Repository";
-import { ResourcePermissionsTestRepository } from "./data/repositories/test/ResourcePermissionsTestRepository";
-import { ResourcePermissionsRepository } from "./domain/repositories/ResourcePermissionsRepository";
 import { GetResourceUserPermissionsUseCase } from "./domain/usecases/GetResourceUserPermissionsUseCase";
 
 export type CompositionRoot = ReturnType<typeof getCompositionRoot>;
@@ -110,7 +107,6 @@ type Repositories = {
     userGroupRepository: UserGroupRepository;
     resourceRepository: ResourceRepository;
     resourceFileRepository: ResourceFileRepository;
-    resourcePermissionsRepository: ResourcePermissionsRepository;
 };
 
 function getCompositionRoot(repositories: Repositories) {
@@ -172,9 +168,7 @@ function getCompositionRoot(repositories: Repositories) {
                 repositories.resourceFileRepository
             ),
             deleteResourceFile: new DeleteResourceFileUseCase(repositories),
-            getPermissions: new GetResourceUserPermissionsUseCase(
-                repositories.resourcePermissionsRepository
-            ),
+            getPermissions: new GetResourceUserPermissionsUseCase(repositories),
         },
     };
 }
@@ -201,7 +195,6 @@ export function getWebappCompositionRoot(api: D2Api) {
         userGroupRepository: new UserGroupD2Repository(api),
         resourceRepository: new ResourceD2Repository(api),
         resourceFileRepository: new ResourceFileD2Repository(api),
-        resourcePermissionsRepository: new ResourcePermissionsD2Repository(api),
     };
 
     return getCompositionRoot(repositories);
@@ -228,7 +221,6 @@ export function getTestCompositionRoot() {
         userGroupRepository: new UserGroupTestRepository(),
         resourceRepository: new ResourceTestRepository(),
         resourceFileRepository: new ResourceFileTestRepository(),
-        resourcePermissionsRepository: new ResourcePermissionsTestRepository(),
     };
 
     return getCompositionRoot(repositories);
